% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeesApproxTMB.R
\name{SCA_LeesApprox}
\alias{SCA_LeesApprox}
\title{SCA with interpolation for Lee's approximation}
\usage{
SCA_LeesApprox(x = 1, Data, SR = c("BH", "Ricker"),
  vulnerability = c("logistic", "dome"), CAA_multiplier = 0,
  CAL_multiplier = 25, ngtg = 3L, max_sd_gtg = 2, I_type = c("B",
  "VB", "SSB"), rescale = "mean1", max_age = Data@MaxAge,
  start = NULL, fix_h = TRUE, fix_F_equilibrium = TRUE,
  fix_omega = TRUE, fix_sigma = FALSE, fix_tau = TRUE,
  early_dev = c("comp_onegen", "comp", "all"), late_dev = "comp50",
  integrate = FALSE, silent = TRUE, opt_hess = FALSE,
  n_restart = ifelse(opt_hess, 0, 1), control = list(iter.max = 2e+05,
  eval.max = 4e+05), inner.control = list(), ...)
}
\description{
SCA with interpolation for Lee's approximation
}
